Moved some defintions to utils.hpp
